backend/api_service/requirements.txt
Flask
psycopg2-binary
google-generativeai
python-dotenv # For local testing only
